refactor(models): split models into github, download, config, logging and re-export in __init__
